+++ mods/BrewLAN_Units/BrewTurrets/hook/lua/ui/help/tooltips.lua
@@ -1,5 +1,5 @@
 Tooltips['activate_weapon'] = {
-      title = "<LOC tooltipbrewturrets0000>Toggle weapon power",
-      description = "<LOC tooltipbrewturrets0001>Enables or disables main weapon power.",
-      keyID = "activate_weapon",
-   }
+    title = "<LOC tooltipbrewturrets0000>Toggle weapon power",
+    description = "<LOC tooltipbrewturrets0001>Enables or disables main weapon power.",
+    keyID = "activate_weapon",
+}